move ServiceError into error module, implement ApiService.check_license()
